Factor operator main into app.Run, add proxyrunner Run helper
Implements changes for issue #5326:
- Move all pre-Run setup from cmd/thv-operator/main.go (scheme registration,
  flag parsing, logger init, signal context, manager construction) into a
  new cmd/thv-operator/app package exposing a single Run() entry point.
- Add additive Run() helper to cmd/thv-proxyrunner/app that absorbs the
  viper env binding, logger init, signal-aware context, and root command
  execution from cmd/thv-proxyrunner/main.go.
- Shrink both main.go files to two-line wrappers that call app.Run(),
  enabling out-of-tree builds to wrap the binaries via a literal
  func main() { app.Run() }.

This is a mechanical code move with no behavior change. All existing
flags, environment variables, exit codes, and signal handling are
preserved unchanged.
